requirements.txt
PyQt6
PyQt6-sip
PyQt6-tools
llama-cpp-python
tk
